+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Terminate commands if not finished after a timeout (default 20s)
 - Show a progress notification display with a cancel button for long running commands.
 - Automatically open a preview on file open if the file type is configured (with customizable position and toggle settings)
+- Respect fileAssociations declaration order to enable custom configuration priorities
 
 ## [0.0.3] - 2026-05-15